lanjun-be/pyproject.toml
[tool.poetry]
name = "lanjun"
version = "0.1.0"
description = ""
authors = ["Spyros Roum <spyros.roum@posteo.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = { extras = ["standard"], version = "^0.20" }
fastapi = "^0.88"
pydantic = "^1.10"
asyncpg = "^0.27"
SQLAlchemy = { extras = ["asyncio"], version = "^1.4" }
python-dotenv = "^0.21"
alembic = "^1.9"
sqlmodel = "^0.0.8"
python-jose = { extras = ["cryptography"], version="^3.3.0" }
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.10"
black = "^22.10"
mypy = "^0.982"
flake8 = "^6.0"
Flake8-pyproject = "^1.2.0"
types-SQLAlchemy = "^1.4"
pytest-alembic = "^0.8.4"
pytest-asyncio = "^0.20.3"
fastapi = { extras = ["test"], version = "^0.88.0" }
types-python-jose = "^3.3.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.isort]
line_length = 99
profile = "black"
skip_glob = ".venv,venv"
float_to_top = true

[tool.black]
line-length = 99
target-version = ["py310"]
exclude = '''
/(
    \.git|venv|.venv
)/
'''

[tool.mypy]
exclude = ".?venv"
check_untyped_defs = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_unreachable = true

[tool.flake8]
max-line-length = 99
extend-ignore = "E203"
